Split preFold() from functionResolution.cpp in to preFold.cpp
